requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
gunicorn>=21.2
python-dotenv>=1.0
Pillow>=10.0
openpyxl>=3.1
xhtml2pdf>=0.2
djangorestframework>=3.14
drf-yasg>=1.21
whitenoise>=6.6
dj-database-url>=2.1
django-unfold>=0.18.0